refill reports

+++ src/NJournals/NJournals.Common/DataEntities/RefillDaySummaryDataEntity.cs
@@ -17,7 +17,7 @@ namespace NJournals.Common.DataEntities
 	public class RefillDaySummaryDataEntity
 	{
 		public virtual int DayID {get;set;}
-		public virtual double TotalSales {get;set;}
+		public virtual decimal TotalSales {get;set;}
 		public virtual int TransCount {get;set;}
 		public virtual DateTime DayStamp {get;set;}
